working on getting an axis in floating layer.
so far it looks like the best bet is to add a second map
type overlay slightly offset from the first.

Subclassing using some kind of control seems unlikely, as
the API limited to positioning a generating dom element to
1 of 12 fixed positions.

It may be possible to get this overlay to listen for
map move for center changing events, and then update the
position, but that seems more expensive than just adding
an overlay that does the right thing.

Next steps: use the same or a second projection, create a
second Image Map?

XXX: draw is called every so often after all.