calculateBounds is used by Map's moveTo.  It's the last piece that performs calculations according to resolution.
After replacing it, moving, lon lat, and layout all appear
correct, but restrictedExtent now fails to work as
expected.

The logic for restrictedExtent is in moveTo, which is a
massive gnarly function.  This instruments the code so
that I can inspect that's going on during the
restrictedExtent logic.  This implementation exactly
matches the implementation provided by openlayers.org.

--HG--
branch : projection